Extract callback tests to separate test case
